requirements.txt
flask
matplotlib
mysql-connector-python
